requirements.txt
sqlite-utils
python-dotenv
requests
tkinterweb
nltk
pandas
matplotlib
fastapi
uvicorn[standard]
beautifulsoup4

